+++ packages/core/src/index.ts
@@ -13,6 +13,7 @@ export * from "./catalog";
 export * from "./chatMessage";
 export * from "./checklists";
 export * from "./co2";
+export * from "./countries";
 export * from "./csv";
 export * from "./csvParse";
 export * from "./damageClaims";